requirements.txt
openai
streamlit
python-dotenv
pypdf
python-dotenv